Updated README since I'm not using the request module for now.
I thought I wanted to use it to redirect the vote form submission. This
is because the form submission is a POST request by law. But I wanted
my poll resource updating to be done via a PUT request, since that is
more RESTful. But in the end, I decided to abandon pure RESTfulness for
easiness.