Switched notify module to temporarily log to a file.
I'm not sure how hard it will be to get SMTP up and running. I may not
have it in time for the demo this wednesday. This log file will be an
easier way to record the lifecycle of a poll. It will also be fairly
straigtforward to switch all places where things are logged to emails.